+++ src/api/definitions.ts
@@ -10,7 +10,7 @@ import {
 import { CurrencyCode } from '../types/currencyCode';
 import { PaymentResult } from '../types/payments';
 import { DataSignature } from '../types/signature';
-import { CreateItemsOrder, ItemTransferResult, Item } from '../types/items';
+import { CreateItemsOrder, ItemTransferResult, Item, ItemTransferAndCreateItemsOrder } from '../types/items';
 
 type PathWithVariable<
 	Prefix extends string,
@@ -52,6 +52,7 @@ export type CloudResponse = {
 	'/v3/itemCreationOrder': CreateItemsOrder;
 	'/v3/itemCreationOrder/createBatch': CreateItemsOrder;
 	'/v3/itemCreationOrder/issueItems': CreateItemsOrder;
+	'/v3/itemCreationOrder/burnAndCreate': ItemTransferAndCreateItemsOrder;
 } & {
 	[K in PathWithVariable<'/v1/connect/wallet/exchangeRate', CurrencyCode, ''>]: ExchangeRate;
 } & {
@@ -62,6 +63,8 @@ export type CloudResponse = {
 	[K in PathWithVariable<'/v3/itemCreationOrder', string, ''>]: CreateItemsOrder;
 } & {
 	[K in PathWithVariable<'/v3/itemCreationOrder', string, '/items'>]: ListResponse<Item>;
+} & {
+	[K in PathWithVariable<'/v3/wallet/items', string, ''>]: Item;
 };
 
 export type CloudEndpoint = keyof CloudResponse;